Moves TestComponent to its own file
